Reworked jsonrpc library a bit
