+++ pyproject.toml
@@ -49,6 +49,7 @@ dependencies = [
 
   # --- Testing dependencies ---
   "coverage==7.6.0",
+  "ipdb==0.13.13",
 
   # --- For notebook development ---
   "jupyter==1.0.0",
